Move safepoints and stack maps tests to the `safepoints` module
