+++ cranelift/frontend/src/frontend.rs
@@ -1206,7 +1206,7 @@ mod tests {
     use alloc::string::ToString;
     use cranelift_codegen::entity::EntityRef;
     use cranelift_codegen::ir::condcodes::IntCC;
-    use cranelift_codegen::ir::{self, types::*, UserFuncName};
+    use cranelift_codegen::ir::{types::*, UserFuncName};
     use cranelift_codegen::ir::{AbiParam, Function, InstBuilder, MemFlags, Signature, Value};
     use cranelift_codegen::isa::{CallConv, TargetFrontendConfig, TargetIsa};
     use cranelift_codegen::settings;
@@ -1940,1173 +1940,5 @@ block0:
                 ))
             );
         }
-    }
-
-    #[test]
-    fn needs_stack_map_and_loop() {
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-        sig.params.push(AbiParam::new(ir::types::I32));
-
-        let mut fn_ctx = FunctionBuilderContext::new();
-        let mut func = Function::with_name_signature(UserFuncName::testcase("sample"), sig);
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-        let signature = builder.func.import_signature(sig);
-        let func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        // Here the value `v1` is technically not live but our single-pass liveness
-        // analysis treats every branch argument to a block as live to avoid
-        // needing to do a fixed-point loop.
-        //
-        //     block0(v0, v1):
-        //       call $foo(v0)
-        //       jump block0(v0, v1)
-        let block0 = builder.create_block();
-        builder.append_block_params_for_function_params(block0);
-        let a = builder.func.dfg.block_params(block0)[0];
-        let b = builder.func.dfg.block_params(block0)[1];
-        builder.declare_value_needs_stack_map(a);
-        builder.declare_value_needs_stack_map(b);
-        builder.switch_to_block(block0);
-        builder.ins().call(func_ref, &[a]);
-        builder.ins().jump(block0, &[a, b]);
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function %sample(i32, i32) system_v {
-    ss0 = explicit_slot 4, align = 4
-    ss1 = explicit_slot 4, align = 4
-    sig0 = (i32) system_v
-    fn0 = colocated u0:0 sig0
-
-block0(v0: i32, v1: i32):
-    stack_store v0, ss0
-    stack_store v1, ss1
-    call fn0(v0), stack_map=[i32 @ ss0+0, i32 @ ss1+0]
-    v2 = stack_load.i32 ss0
-    v3 = stack_load.i32 ss1
-    jump block0(v2, v3)
-}            "#
-                .trim()
-        );
-    }
-
-    #[test]
-    fn needs_stack_map_simple() {
-        let sig = Signature::new(CallConv::SystemV);
-
-        let mut fn_ctx = FunctionBuilderContext::new();
-        let mut func = Function::with_name_signature(UserFuncName::testcase("sample"), sig);
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-        let signature = builder.func.import_signature(sig);
-        let func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        // At each `call` we are losing one more value as no longer live, so
-        // each stack map should be one smaller than the last. `v3` is never
-        // live across a safepoint, so should never appear in a stack map. Note
-        // that a value that is an argument to the call, but is not live after
-        // the call, should not appear in the stack map. This is why `v0`
-        // appears in the first call's stack map, but not the second call's
-        // stack map.
-        //
-        //     block0:
-        //       v0 = needs stack map
-        //       v1 = needs stack map
-        //       v2 = needs stack map
-        //       v3 = needs stack map
-        //       call $foo(v3)
-        //       call $foo(v0)
-        //       call $foo(v1)
-        //       call $foo(v2)
-        //       return
-        let block0 = builder.create_block();
-        builder.append_block_params_for_function_params(block0);
-        builder.switch_to_block(block0);
-        let v0 = builder.ins().iconst(ir::types::I32, 0);
-        builder.declare_value_needs_stack_map(v0);
-        let v1 = builder.ins().iconst(ir::types::I32, 1);
-        builder.declare_value_needs_stack_map(v1);
-        let v2 = builder.ins().iconst(ir::types::I32, 2);
-        builder.declare_value_needs_stack_map(v2);
-        let v3 = builder.ins().iconst(ir::types::I32, 3);
-        builder.declare_value_needs_stack_map(v3);
-        builder.ins().call(func_ref, &[v3]);
-        builder.ins().call(func_ref, &[v0]);
-        builder.ins().call(func_ref, &[v1]);
-        builder.ins().call(func_ref, &[v2]);
-        builder.ins().return_(&[]);
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function %sample() system_v {
-    ss0 = explicit_slot 4, align = 4
-    ss1 = explicit_slot 4, align = 4
-    ss2 = explicit_slot 4, align = 4
-    sig0 = (i32) system_v
-    fn0 = colocated u0:0 sig0
-
-block0:
-    v0 = iconst.i32 0
-    stack_store v0, ss2  ; v0 = 0
-    v1 = iconst.i32 1
-    stack_store v1, ss1  ; v1 = 1
-    v2 = iconst.i32 2
-    stack_store v2, ss0  ; v2 = 2
-    v3 = iconst.i32 3
-    call fn0(v3), stack_map=[i32 @ ss2+0, i32 @ ss1+0, i32 @ ss0+0]  ; v3 = 3
-    v4 = stack_load.i32 ss2
-    call fn0(v4), stack_map=[i32 @ ss1+0, i32 @ ss0+0]
-    v5 = stack_load.i32 ss1
-    call fn0(v5), stack_map=[i32 @ ss0+0]
-    v6 = stack_load.i32 ss0
-    call fn0(v6)
-    return
-}
-            "#
-            .trim()
-        );
-    }
-
-    #[test]
-    fn needs_stack_map_and_post_order_early_return() {
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-
-        let mut fn_ctx = FunctionBuilderContext::new();
-        let mut func = Function::with_name_signature(UserFuncName::testcase("sample"), sig);
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let signature = builder
-            .func
-            .import_signature(Signature::new(CallConv::SystemV));
-        let func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        // Here we rely on the post-order to make sure that we never visit block
-        // 4 and add `v1` to our live set, then visit block 2 and add `v1` to
-        // its stack map even though `v1` is not in scope. Thanksfully, that
-        // sequence is impossible because it would be an invalid post-order
-        // traversal. The only valid post-order traversals are [3, 1, 2, 0] and
-        // [2, 3, 1, 0].
-        //
-        //     block0(v0):
-        //       brif v0, block1, block2
-        //
-        //     block1:
-        //       <stuff>
-        //       v1 = get some gc ref
-        //       jump block3
-        //
-        //     block2:
-        //       call $needs_safepoint_accidentally
-        //       return
-        //
-        //     block3:
-        //       stuff keeping v1 live
-        //       return
-        let block0 = builder.create_block();
-        let block1 = builder.create_block();
-        let block2 = builder.create_block();
-        let block3 = builder.create_block();
-        builder.append_block_params_for_function_params(block0);
-
-        builder.switch_to_block(block0);
-        let v0 = builder.func.dfg.block_params(block0)[0];
-        builder.ins().brif(v0, block1, &[], block2, &[]);
-
-        builder.switch_to_block(block1);
-        let v1 = builder.ins().iconst(ir::types::I64, 0x12345678);
-        builder.declare_value_needs_stack_map(v1);
-        builder.ins().jump(block3, &[]);
-
-        builder.switch_to_block(block2);
-        builder.ins().call(func_ref, &[]);
-        builder.ins().return_(&[]);
-
-        builder.switch_to_block(block3);
-        // NB: Our simplistic liveness analysis conservatively treats any use of
-        // a value as keeping it live, regardless if the use has side effects or
-        // is otherwise itself live, so an `iadd_imm` suffices to keep `v1` live
-        // here.
-        builder.ins().iadd_imm(v1, 0);
-        builder.ins().return_(&[]);
-
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function %sample(i32) system_v {
-    sig0 = () system_v
-    fn0 = colocated u0:0 sig0
-
-block0(v0: i32):
-    brif v0, block1, block2
-
-block1:
-    v1 = iconst.i64 0x1234_5678
-    jump block3
-
-block2:
-    call fn0()
-    return
-
-block3:
-    v2 = iadd_imm.i64 v1, 0  ; v1 = 0x1234_5678
-    return
-}
-            "#
-            .trim()
-        );
-    }
-
-    #[test]
-    fn needs_stack_map_conditional_branches_and_liveness() {
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-
-        let mut fn_ctx = FunctionBuilderContext::new();
-        let mut func = Function::with_name_signature(UserFuncName::testcase("sample"), sig);
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let signature = builder
-            .func
-            .import_signature(Signature::new(CallConv::SystemV));
-        let func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        // Depending on which post-order traversal we take, we might consider
-        // `v1` live inside `block1` and emit unnecessary safepoint
-        // spills. That's not great, but ultimately fine, we are trading away
-        // precision for a single-pass analysis.
-        //
-        //     block0(v0):
-        //       v1 = needs stack map
-        //       brif v0, block1, block2
-        //
-        //     block1:
-        //       call $foo()
-        //       return
-        //
-        //     block2:
-        //       keep v1 alive
-        //       return
-        let block0 = builder.create_block();
-        let block1 = builder.create_block();
-        let block2 = builder.create_block();
-        builder.append_block_params_for_function_params(block0);
-
-        builder.switch_to_block(block0);
-        let v0 = builder.func.dfg.block_params(block0)[0];
-        let v1 = builder.ins().iconst(ir::types::I64, 0x12345678);
-        builder.declare_value_needs_stack_map(v1);
-        builder.ins().brif(v0, block1, &[], block2, &[]);
-
-        builder.switch_to_block(block1);
-        builder.ins().call(func_ref, &[]);
-        builder.ins().return_(&[]);
-
-        builder.switch_to_block(block2);
-        // NB: Our simplistic liveness analysis conservatively treats any use of
-        // a value as keeping it live, regardless if the use has side effects or
-        // is otherwise itself live, so an `iadd_imm` suffices to keep `v1` live
-        // here.
-        builder.ins().iadd_imm(v1, 0);
-        builder.ins().return_(&[]);
-
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function %sample(i32) system_v {
-    sig0 = () system_v
-    fn0 = colocated u0:0 sig0
-
-block0(v0: i32):
-    v1 = iconst.i64 0x1234_5678
-    brif v0, block1, block2
-
-block1:
-    call fn0()
-    return
-
-block2:
-    v2 = iadd_imm.i64 v1, 0  ; v1 = 0x1234_5678
-    return
-}
-            "#
-            .trim()
-        );
-
-        // Now Do the same test but with block 1 and 2 swapped so that we
-        // exercise what we are trying to exercise, regardless of which
-        // post-order traversal we happen to take.
-        func.clear();
-        fn_ctx.clear();
-
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-
-        func.signature = sig;
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let signature = builder
-            .func
-            .import_signature(Signature::new(CallConv::SystemV));
-        let func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        let block0 = builder.create_block();
-        let block1 = builder.create_block();
-        let block2 = builder.create_block();
-        builder.append_block_params_for_function_params(block0);
-
-        builder.switch_to_block(block0);
-        let v0 = builder.func.dfg.block_params(block0)[0];
-        let v1 = builder.ins().iconst(ir::types::I64, 0x12345678);
-        builder.declare_value_needs_stack_map(v1);
-        builder.ins().brif(v0, block1, &[], block2, &[]);
-
-        builder.switch_to_block(block1);
-        builder.ins().iadd_imm(v1, 0);
-        builder.ins().return_(&[]);
-
-        builder.switch_to_block(block2);
-        builder.ins().call(func_ref, &[]);
-        builder.ins().return_(&[]);
-
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function u0:0(i32) system_v {
-    ss0 = explicit_slot 8, align = 8
-    sig0 = () system_v
-    fn0 = colocated u0:0 sig0
-
-block0(v0: i32):
-    v1 = iconst.i64 0x1234_5678
-    stack_store v1, ss0  ; v1 = 0x1234_5678
-    brif v0, block1, block2
-
-block1:
-    v3 = stack_load.i64 ss0
-    v2 = iadd_imm v3, 0
-    return
-
-block2:
-    call fn0(), stack_map=[i64 @ ss0+0]
-    return
-}
-            "#
-            .trim()
-        );
-    }
-
-    #[test]
-    fn needs_stack_map_and_tail_calls() {
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-
-        let mut fn_ctx = FunctionBuilderContext::new();
-        let mut func = Function::with_name_signature(UserFuncName::testcase("sample"), sig);
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let signature = builder
-            .func
-            .import_signature(Signature::new(CallConv::SystemV));
-        let func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        // Depending on which post-order traversal we take, we might consider
-        // `v1` live inside `block1`. But nothing is live after a tail call so
-        // we shouldn't spill `v1` either way here.
-        //
-        //     block0(v0):
-        //       v1 = needs stack map
-        //       brif v0, block1, block2
-        //
-        //     block1:
-        //       return_call $foo()
-        //
-        //     block2:
-        //       keep v1 alive
-        //       return
-        let block0 = builder.create_block();
-        let block1 = builder.create_block();
-        let block2 = builder.create_block();
-        builder.append_block_params_for_function_params(block0);
-
-        builder.switch_to_block(block0);
-        let v0 = builder.func.dfg.block_params(block0)[0];
-        let v1 = builder.ins().iconst(ir::types::I64, 0x12345678);
-        builder.declare_value_needs_stack_map(v1);
-        builder.ins().brif(v0, block1, &[], block2, &[]);
-
-        builder.switch_to_block(block1);
-        builder.ins().return_call(func_ref, &[]);
-
-        builder.switch_to_block(block2);
-        // NB: Our simplistic liveness analysis conservatively treats any use of
-        // a value as keeping it live, regardless if the use has side effects or
-        // is otherwise itself live, so an `iadd_imm` suffices to keep `v1` live
-        // here.
-        builder.ins().iadd_imm(v1, 0);
-        builder.ins().return_(&[]);
-
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function %sample(i32) system_v {
-    sig0 = () system_v
-    fn0 = colocated u0:0 sig0
-
-block0(v0: i32):
-    v1 = iconst.i64 0x1234_5678
-    brif v0, block1, block2
-
-block1:
-    return_call fn0()
-
-block2:
-    v2 = iadd_imm.i64 v1, 0  ; v1 = 0x1234_5678
-    return
-}
-            "#
-            .trim()
-        );
-
-        // Do the same test but with block 1 and 2 swapped so that we exercise
-        // what we are trying to exercise, regardless of which post-order
-        // traversal we happen to take.
-        func.clear();
-        fn_ctx.clear();
-
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-        func.signature = sig;
-
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let signature = builder
-            .func
-            .import_signature(Signature::new(CallConv::SystemV));
-        let func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        let block0 = builder.create_block();
-        let block1 = builder.create_block();
-        let block2 = builder.create_block();
-        builder.append_block_params_for_function_params(block0);
-
-        builder.switch_to_block(block0);
-        let v0 = builder.func.dfg.block_params(block0)[0];
-        let v1 = builder.ins().iconst(ir::types::I64, 0x12345678);
-        builder.declare_value_needs_stack_map(v1);
-        builder.ins().brif(v0, block1, &[], block2, &[]);
-
-        builder.switch_to_block(block1);
-        builder.ins().iadd_imm(v1, 0);
-        builder.ins().return_(&[]);
-
-        builder.switch_to_block(block2);
-        builder.ins().return_call(func_ref, &[]);
-
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function u0:0(i32) system_v {
-    sig0 = () system_v
-    fn0 = colocated u0:0 sig0
-
-block0(v0: i32):
-    v1 = iconst.i64 0x1234_5678
-    brif v0, block1, block2
-
-block1:
-    v2 = iadd_imm.i64 v1, 0  ; v1 = 0x1234_5678
-    return
-
-block2:
-    return_call fn0()
-}
-            "#
-            .trim()
-        );
-    }
-
-    #[test]
-    fn needs_stack_map_and_cfg_diamond() {
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-
-        let mut fn_ctx = FunctionBuilderContext::new();
-        let mut func = Function::with_name_signature(UserFuncName::testcase("sample"), sig);
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let signature = builder
-            .func
-            .import_signature(Signature::new(CallConv::SystemV));
-        let func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        // Create an if/else CFG diamond that and check that various things get
-        // spilled as needed.
-        //
-        //     block0(v0):
-        //       brif v0, block1, block2
-        //
-        //     block1:
-        //       v1 = needs stack map
-        //       v2 = needs stack map
-        //       call $foo()
-        //       jump block3(v1, v2)
-        //
-        //     block2:
-        //       v3 = needs stack map
-        //       v4 = needs stack map
-        //       call $foo()
-        //       jump block3(v3, v3)  ;; Note: v4 is not live
-        //
-        //     block3(v5, v6):
-        //       call $foo()
-        //       keep v5 alive, but not v6
-        let block0 = builder.create_block();
-        let block1 = builder.create_block();
-        let block2 = builder.create_block();
-        let block3 = builder.create_block();
-        builder.append_block_params_for_function_params(block0);
-
-        builder.switch_to_block(block0);
-        let v0 = builder.func.dfg.block_params(block0)[0];
-        builder.ins().brif(v0, block1, &[], block2, &[]);
-
-        builder.switch_to_block(block1);
-        let v1 = builder.ins().iconst(ir::types::I64, 1);
-        builder.declare_value_needs_stack_map(v1);
-        let v2 = builder.ins().iconst(ir::types::I64, 2);
-        builder.declare_value_needs_stack_map(v2);
-        builder.ins().call(func_ref, &[]);
-        builder.ins().jump(block3, &[v1, v2]);
-
-        builder.switch_to_block(block2);
-        let v3 = builder.ins().iconst(ir::types::I64, 3);
-        builder.declare_value_needs_stack_map(v3);
-        let v4 = builder.ins().iconst(ir::types::I64, 4);
-        builder.declare_value_needs_stack_map(v4);
-        builder.ins().call(func_ref, &[]);
-        builder.ins().jump(block3, &[v3, v3]);
-
-        builder.switch_to_block(block3);
-        builder.append_block_param(block3, ir::types::I64);
-        builder.append_block_param(block3, ir::types::I64);
-        builder.ins().call(func_ref, &[]);
-        // NB: Our simplistic liveness analysis conservatively treats any use of
-        // a value as keeping it live, regardless if the use has side effects or
-        // is otherwise itself live, so an `iadd_imm` suffices to keep `v1` live
-        // here.
-        builder.ins().iadd_imm(v1, 0);
-        builder.ins().return_(&[]);
-
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function %sample(i32) system_v {
-    ss0 = explicit_slot 8, align = 8
-    ss1 = explicit_slot 8, align = 8
-    sig0 = () system_v
-    fn0 = colocated u0:0 sig0
-
-block0(v0: i32):
-    brif v0, block1, block2
-
-block1:
-    v1 = iconst.i64 1
-    stack_store v1, ss0  ; v1 = 1
-    v2 = iconst.i64 2
-    stack_store v2, ss1  ; v2 = 2
-    call fn0(), stack_map=[i64 @ ss0+0, i64 @ ss1+0]
-    v8 = stack_load.i64 ss0
-    v9 = stack_load.i64 ss1
-    jump block3(v8, v9)
-
-block2:
-    v3 = iconst.i64 3
-    stack_store v3, ss0  ; v3 = 3
-    v4 = iconst.i64 4
-    call fn0(), stack_map=[i64 @ ss0+0]
-    v10 = stack_load.i64 ss0
-    v11 = stack_load.i64 ss0
-    jump block3(v10, v11)
-
-block3(v5: i64, v6: i64):
-    call fn0(), stack_map=[i64 @ ss0+0]
-    v12 = stack_load.i64 ss0
-    v7 = iadd_imm v12, 0
-    return
-}
-            "#
-            .trim()
-        );
-    }
-
-    #[test]
-    fn needs_stack_map_and_heterogeneous_types() {
-        let mut sig = Signature::new(CallConv::SystemV);
-        for ty in [
-            ir::types::I8,
-            ir::types::I16,
-            ir::types::I32,
-            ir::types::I64,
-            ir::types::I128,
-            ir::types::F32,
-            ir::types::F64,
-            ir::types::I8X16,
-            ir::types::I16X8,
-        ] {
-            sig.params.push(AbiParam::new(ty));
-            sig.returns.push(AbiParam::new(ty));
-        }
-
-        let mut fn_ctx = FunctionBuilderContext::new();
-        let mut func = Function::with_name_signature(UserFuncName::testcase("sample"), sig);
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let signature = builder
-            .func
-            .import_signature(Signature::new(CallConv::SystemV));
-        let func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        // Test that we support stack maps of heterogeneous types and properly
-        // coalesce types into stack slots based on their size.
-        //
-        //     block0(v0, v1, v2, v3, v4, v5, v6, v7, v8):
-        //       call $foo()
-        //       return v0, v1, v2, v3, v4, v5, v6, v7, v8
-        let block0 = builder.create_block();
-        builder.append_block_params_for_function_params(block0);
-
-        builder.switch_to_block(block0);
-        let params = builder.func.dfg.block_params(block0).to_vec();
-        for val in &params {
-            builder.declare_value_needs_stack_map(*val);
-        }
-        builder.ins().call(func_ref, &[]);
-        builder.ins().return_(&params);
-
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function %sample(i8, i16, i32, i64, i128, f32, f64, i8x16, i16x8) -> i8, i16, i32, i64, i128, f32, f64, i8x16, i16x8 system_v {
-    ss0 = explicit_slot 1
-    ss1 = explicit_slot 2, align = 2
-    ss2 = explicit_slot 4, align = 4
-    ss3 = explicit_slot 8, align = 8
-    ss4 = explicit_slot 16, align = 16
-    ss5 = explicit_slot 4, align = 4
-    ss6 = explicit_slot 8, align = 8
-    ss7 = explicit_slot 16, align = 16
-    ss8 = explicit_slot 16, align = 16
-    sig0 = () system_v
-    fn0 = colocated u0:0 sig0
-
-block0(v0: i8, v1: i16, v2: i32, v3: i64, v4: i128, v5: f32, v6: f64, v7: i8x16, v8: i16x8):
-    stack_store v0, ss0
-    stack_store v1, ss1
-    stack_store v2, ss2
-    stack_store v3, ss3
-    stack_store v4, ss4
-    stack_store v5, ss5
-    stack_store v6, ss6
-    stack_store v7, ss7
-    stack_store v8, ss8
-    call fn0(), stack_map=[i8 @ ss0+0, i16 @ ss1+0, i32 @ ss2+0, i64 @ ss3+0, i128 @ ss4+0, f32 @ ss5+0, f64 @ ss6+0, i8x16 @ ss7+0, i16x8 @ ss8+0]
-    v9 = stack_load.i8 ss0
-    v10 = stack_load.i16 ss1
-    v11 = stack_load.i32 ss2
-    v12 = stack_load.i64 ss3
-    v13 = stack_load.i128 ss4
-    v14 = stack_load.f32 ss5
-    v15 = stack_load.f64 ss6
-    v16 = stack_load.i8x16 ss7
-    v17 = stack_load.i16x8 ss8
-    return v9, v10, v11, v12, v13, v14, v15, v16, v17
-}
-            "#
-            .trim()
-        );
-    }
-
-    #[test]
-    fn series_of_non_overlapping_live_ranges_needs_stack_map() {
-        let sig = Signature::new(CallConv::SystemV);
-
-        let mut fn_ctx = FunctionBuilderContext::new();
-        let mut func = Function::with_name_signature(UserFuncName::testcase("sample"), sig);
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let signature = builder
-            .func
-            .import_signature(Signature::new(CallConv::SystemV));
-        let foo_func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 1,
-            });
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-        let signature = builder.func.import_signature(sig);
-        let consume_func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        // Create a series of needs-stack-map values that do not have
-        // overlapping live ranges, but which do appear in stack maps for calls
-        // to `$foo`:
-        //
-        //     block0:
-        //       v0 = needs stack map
-        //       call $foo()
-        //       call consume(v0)
-        //       v1 = needs stack map
-        //       call $foo()
-        //       call consume(v1)
-        //       v2 = needs stack map
-        //       call $foo()
-        //       call consume(v2)
-        //       v3 = needs stack map
-        //       call $foo()
-        //       call consume(v3)
-        //       return
-        let block0 = builder.create_block();
-        builder.append_block_params_for_function_params(block0);
-        builder.switch_to_block(block0);
-        let v0 = builder.ins().iconst(ir::types::I32, 0);
-        builder.declare_value_needs_stack_map(v0);
-        builder.ins().call(foo_func_ref, &[]);
-        builder.ins().call(consume_func_ref, &[v0]);
-        let v1 = builder.ins().iconst(ir::types::I32, 1);
-        builder.declare_value_needs_stack_map(v1);
-        builder.ins().call(foo_func_ref, &[]);
-        builder.ins().call(consume_func_ref, &[v1]);
-        let v2 = builder.ins().iconst(ir::types::I32, 2);
-        builder.declare_value_needs_stack_map(v2);
-        builder.ins().call(foo_func_ref, &[]);
-        builder.ins().call(consume_func_ref, &[v2]);
-        let v3 = builder.ins().iconst(ir::types::I32, 3);
-        builder.declare_value_needs_stack_map(v3);
-        builder.ins().call(foo_func_ref, &[]);
-        builder.ins().call(consume_func_ref, &[v3]);
-        builder.ins().return_(&[]);
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function %sample() system_v {
-    ss0 = explicit_slot 4, align = 4
-    sig0 = () system_v
-    sig1 = (i32) system_v
-    fn0 = colocated u0:0 sig0
-    fn1 = colocated u0:1 sig1
-
-block0:
-    v0 = iconst.i32 0
-    stack_store v0, ss0  ; v0 = 0
-    call fn0(), stack_map=[i32 @ ss0+0]
-    v4 = stack_load.i32 ss0
-    call fn1(v4)
-    v1 = iconst.i32 1
-    stack_store v1, ss0  ; v1 = 1
-    call fn0(), stack_map=[i32 @ ss0+0]
-    v5 = stack_load.i32 ss0
-    call fn1(v5)
-    v2 = iconst.i32 2
-    stack_store v2, ss0  ; v2 = 2
-    call fn0(), stack_map=[i32 @ ss0+0]
-    v6 = stack_load.i32 ss0
-    call fn1(v6)
-    v3 = iconst.i32 3
-    stack_store v3, ss0  ; v3 = 3
-    call fn0(), stack_map=[i32 @ ss0+0]
-    v7 = stack_load.i32 ss0
-    call fn1(v7)
-    return
-}
-            "#
-            .trim()
-        );
-    }
-
-    #[test]
-    fn vars_block_params_and_needs_stack_map() {
-        let _ = env_logger::try_init();
-
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-        sig.returns.push(AbiParam::new(ir::types::I32));
-
-        let mut fn_ctx = FunctionBuilderContext::new();
-        let mut func = Function::with_name_signature(UserFuncName::testcase("sample"), sig);
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params.push(AbiParam::new(ir::types::I32));
-        let signature = builder.func.import_signature(sig);
-        let func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        // Use a variable, create a control flow diamond so that the variable
-        // forces a block parameter on the control join point, and make sure
-        // that we get stack maps for all the appropriate uses of the variable
-        // in all blocks, as well as that we are reusing stack slots for each of
-        // the values.
-        //
-        //                        block0:
-        //                          x := needs stack map
-        //                          call $foo(x)
-        //                          br_if v0, block1, block2
-        //
-        //
-        //     block1:                                     block2:
-        //       call $foo(x)                                call $foo(x)
-        //       call $foo(x)                                call $foo(x)
-        //       x := new needs stack map                    x := new needs stack map
-        //       call $foo(x)                                call $foo(x)
-        //       jump block3                                 jump block3
-        //
-        //
-        //                        block3:
-        //                          call $foo(x)
-        //                          return x
-
-        let x = Variable::from_u32(0);
-        builder.declare_var(x, ir::types::I32);
-        builder.declare_var_needs_stack_map(x);
-
-        let block0 = builder.create_block();
-        let block1 = builder.create_block();
-        let block2 = builder.create_block();
-        let block3 = builder.create_block();
-
-        builder.append_block_params_for_function_params(block0);
-        builder.switch_to_block(block0);
-        let v0 = builder.func.dfg.block_params(block0)[0];
-        let val = builder.ins().iconst(ir::types::I32, 42);
-        builder.def_var(x, val);
-        {
-            let x = builder.use_var(x);
-            builder.ins().call(func_ref, &[x]);
-        }
-        builder.ins().brif(v0, block1, &[], block2, &[]);
-
-        builder.switch_to_block(block1);
-        {
-            let x = builder.use_var(x);
-            builder.ins().call(func_ref, &[x]);
-            builder.ins().call(func_ref, &[x]);
-        }
-        let val = builder.ins().iconst(ir::types::I32, 36);
-        builder.def_var(x, val);
-        {
-            let x = builder.use_var(x);
-            builder.ins().call(func_ref, &[x]);
-        }
-        builder.ins().jump(block3, &[]);
-
-        builder.switch_to_block(block2);
-        {
-            let x = builder.use_var(x);
-            builder.ins().call(func_ref, &[x]);
-            builder.ins().call(func_ref, &[x]);
-        }
-        let val = builder.ins().iconst(ir::types::I32, 36);
-        builder.def_var(x, val);
-        {
-            let x = builder.use_var(x);
-            builder.ins().call(func_ref, &[x]);
-        }
-        builder.ins().jump(block3, &[]);
-
-        builder.switch_to_block(block3);
-        let x = builder.use_var(x);
-        builder.ins().call(func_ref, &[x]);
-        builder.ins().return_(&[x]);
-
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-
-        // Because our liveness analysis is very simple, and visit blocks in the
-        // order 3->1->2->0, we see uses of `v2` in block1 and mark it live
-        // across all of block2 because we haven't reached the def in block0
-        // yet, even though it isn't technically live out of block2, only live
-        // in. This means that it shows up in the stack map for block2's second
-        // call to `foo()` when it technically needn't, and additionally means
-        // that we have two stack slots instead of a single one below. This
-        // could all be improved and cleaned up by improving the liveness
-        // analysis.
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function %sample(i32) -> i32 system_v {
-    ss0 = explicit_slot 4, align = 4
-    ss1 = explicit_slot 4, align = 4
-    sig0 = (i32) system_v
-    fn0 = colocated u0:0 sig0
-
-block0(v0: i32):
-    v1 = iconst.i32 42
-    v2 -> v1
-    v4 -> v1
-    stack_store v1, ss0  ; v1 = 42
-    v7 = stack_load.i32 ss0
-    call fn0(v7), stack_map=[i32 @ ss0+0]
-    brif v0, block1, block2
-
-block1:
-    call fn0(v2), stack_map=[i32 @ ss0+0]  ; v2 = 42
-    call fn0(v2)  ; v2 = 42
-    v3 = iconst.i32 36
-    stack_store v3, ss0  ; v3 = 36
-    v8 = stack_load.i32 ss0
-    call fn0(v8), stack_map=[i32 @ ss0+0]
-    v9 = stack_load.i32 ss0
-    jump block3(v9)
-
-block2:
-    call fn0(v4), stack_map=[i32 @ ss0+0]  ; v4 = 42
-    call fn0(v4), stack_map=[i32 @ ss0+0]  ; v4 = 42
-    v5 = iconst.i32 36
-    stack_store v5, ss1  ; v5 = 36
-    v10 = stack_load.i32 ss1
-    call fn0(v10), stack_map=[i32 @ ss0+0, i32 @ ss1+0]
-    v11 = stack_load.i32 ss1
-    jump block3(v11)
-
-block3(v6: i32):
-    stack_store v6, ss0
-    call fn0(v6), stack_map=[i32 @ ss0+0]
-    v12 = stack_load.i32 ss0
-    return v12
-}
-            "#
-            .trim()
-        );
-    }
-
-    #[test]
-    fn var_needs_stack_map() {
-        let mut sig = Signature::new(CallConv::SystemV);
-        sig.params
-            .push(AbiParam::new(cranelift_codegen::ir::types::I32));
-        sig.returns
-            .push(AbiParam::new(cranelift_codegen::ir::types::I32));
-
-        let mut fn_ctx = FunctionBuilderContext::new();
-        let mut func = Function::with_name_signature(UserFuncName::testcase("sample"), sig);
-        let mut builder = FunctionBuilder::new(&mut func, &mut fn_ctx);
-
-        let var = Variable::from_u32(0);
-        builder.declare_var(var, cranelift_codegen::ir::types::I32);
-        builder.declare_var_needs_stack_map(var);
-
-        let name = builder
-            .func
-            .declare_imported_user_function(ir::UserExternalName {
-                namespace: 0,
-                index: 0,
-            });
-        let signature = builder
-            .func
-            .import_signature(Signature::new(CallConv::SystemV));
-        let func_ref = builder.import_function(ir::ExtFuncData {
-            name: ir::ExternalName::user(name),
-            signature,
-            colocated: true,
-        });
-
-        let block0 = builder.create_block();
-        builder.append_block_params_for_function_params(block0);
-        builder.switch_to_block(block0);
-
-        let arg = builder.func.dfg.block_params(block0)[0];
-        builder.def_var(var, arg);
-
-        builder.ins().call(func_ref, &[]);
-
-        let val = builder.use_var(var);
-        builder.ins().return_(&[val]);
-
-        builder.seal_all_blocks();
-        builder.finalize();
-
-        eprintln!("Actual = {}", func.display());
-        assert_eq!(
-            func.display().to_string().trim(),
-            r#"
-function %sample(i32) -> i32 system_v {
-    ss0 = explicit_slot 4, align = 4
-    sig0 = () system_v
-    fn0 = colocated u0:0 sig0
-
-block0(v0: i32):
-    stack_store v0, ss0
-    call fn0(), stack_map=[i32 @ ss0+0]
-    v1 = stack_load.i32 ss0
-    return v1
-}
-            "#
-            .trim()
-        );
     }
 }
